PhotosynTuzet {plantecophys}	R Documentation

Coupled leaf gas exchange model with Tuzet stomatal conductance

Description

An implementation of the coupled photosynthesis - stomatal conductance model, using the Tuzet et al. (2003) model of stomatal conductance. Accepts all arguments of Photosyn (except gsmodel, of course).

Usage

PhotosynTuzet(g1 = 8, Ca = 400, psis = 0, kl = 2, sf = 3,
  psif = -2, ...)

Arguments

g1
The slope parameter. Note that the default value should be much higher than that used in the Medlyn et al. (2011) model to give comparable predictions.

Ca
Atmospheric CO2 concentration.

psis
Soil water potential (MPa). Note that soil-to-root hydraulic conductance is not implemented.

kl
Leaf-specific hydraulic conductance (mmol m-2 s-1 MPa-1)

sf
Shape parameter (-) of sigmoidal function of leaf water potential (see Tuzet et al. 2003)

psif
Leaf water potential at which stomatal conductance is 50% of maximum (MPa).

...
All other arguments in Photosyn

[Package plantecophys version 1.4-4 Index]