findCiTransition {plantecophys}	R Documentation

Calculate transition points for fitted A-Ci curves

Description

Calculates the Ci at the transition points between Ac & Aj (point 1), and Aj and Ap (point 2). The latter is not NA only when TPU was estimated (and estimable), see fitaci, argument fitTPU.

Usage

findCiTransition(object, ...)

Arguments

object
Either an object returned by fitaci, or a copy of the Photosyn function.

...
Further arguments passed to the Photosyn function.

Details

This function is also used by fitaci, the results are stored in elements Ci_transition and Ci_transition2.

[Package plantecophys version 1.4-4 Index]