fitBBs {plantecophys}	R Documentation

Fit Ball-Berry type models of stomatal conductance to many groups at once

Description

A batch utility for the fitBB function, to fit the model for each group in a dataframe.

Usage

fitBBs(data, group, ...)

Arguments

data
Input dataframe, containing all variables needed to fit the model.

group
Name of the grouping variable in the dataframe (quoted), the model will be fit for each group defined by this variable.

...
Further parameters passed to fitBB, see there for a full description.

Examples

## Not run: 
# If you have a factor variable in your dataset called 'species', and you
# want to fit the Ball-Berry model for each of the species:
myfits <- fitBBs(mydata, "species", model="BallBerry")

# A dataframe with coefficients is returned by coef()
coef(myfits)

## End(Not run)

[Package plantecophys version 1.4-4 Index]