plantecophys

plantecophys

Details

The DESCRIPTION file: plantecophys plantecophys

The following functions are the main tools in plantecophys:

fitaci (and fitaci) fits A-Ci curves to data.

Photosyn can be used to simulate A-Ci curves (or Aci), and simulate from a coupled leaf gas exchange model.

fitBB fits Ball-Berry-type stomatal conductance models to data.

FARAO is an implementation of a numeric solution to Cowan-Farquhar optimization of stomatal conductance.

RHtoVPD converts relative humidity to vapour pressure deficit (and more similar functions on that help page).

The package also includes the following example datasets:

acidata1 A dataset with a single A-Ci curve.

manyacidat A dataset with many A-Ci curves.

References

Duursma, R.A., 2015. Plantecophys - An R Package for Analysing and Modelling Leaf Gas Exchange Data. PLoS ONE 10, e0143346. doi:10.1371/journal.pone.0143346