Conversions between relative humidity, vapour pressure deficit and dewpoint

A collection of functions to convert between relative humidity (RH) (%), vapour pressure deficit (VPD) (kPa), dew point temperature, and leaf- or air temperature-based VPD or RH. To convert from relative humidity to VPD, use the RHtoVPD function, use VPDtoRH for the other way around. The water vapor saturation pressure is calculated with esat. Use DewtoVPD to convert from dewpoint temperature to VPD. The functions VPDleafToAir and VPDairToLeaf convert VPD from a leaf temperature to an air-temperature basis and vice versa. The functions RHleafToAir a RHairToLeaf do the same for relative humidity.

RHtoVPD(RH, TdegC, Pa = 101)

VPDtoRH(VPD, TdegC, Pa = 101)

esat(TdegC, Pa = 101)

VPDtoDew(VPD, TdegC, Pa = 101)

DewtoVPD(Tdew, TdegC, Pa = 101)

VPDleafToAir(VPD, Tleaf, Tair, Pa = 101)

VPDairToLeaf(VPD, Tair, Tleaf, Pa = 101)

RHleafToAir(RH, Tleaf, Tair, Pa = 101)

RHairToLeaf(RH, Tair, Tleaf, Pa = 101)

Arguments

RH
Relative humidity (%)

TdegC
Temperature (degrees C) (either leaf or air)

Pa
Atmospheric pressure (kPa)

VPD
Vapour pressure deficit (kPa)

Tdew
Dewpoint temperature (degrees C)

Tleaf
Leaf temperature (degrees C)

Tair
Air temperature (degrees C)

Details

The function describing saturated vapor pressure with temperature is taken from Jones (1992). All other calculations follow directly from the standard definitions, for which Jones (1992) may also be consulted.

References

Jones, H.G. 1992. Plants and microclimate: a quantitative approach to environmental plant physiology. 2nd Edition., 2nd Edn. Cambridge University Press, Cambridge. 428 p.